+++ user.h
@@ -24,7 +24,7 @@ char* sbrk(int);
 int sleep(int);
 int uptime(void);
 int symlink(char*, char*);
-int readlink(char*, char*, uint);
+int readlink(char*, char*, size_t);
 int ftag(int, const char*, const char*);	//@Added: system call task 3
 int funtag(int, const char*);				//@Added: system call task 3
 int gettag(int, const char*, const char*);	//@Added: system call task 3
